Fixed menuitems.cpp:3899:93: error: no matching function for call to ‘min(unsigned int, std::vector<sSavedReportMessage>::size_type)’
min template requires both arguments to be the same type. Im not sure whether its fine to simply cast here but if I got this right we are requesting the size of an std::vector here and this is an unsigned int according to cpp reference.

git-svn-id: svn://maxr.org/MAX/trunk@3408 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ menuitems.cpp
@@ -3896,7 +3896,7 @@ void cMenuReportsScreen::drawReportsScreen()
 	SDL_Rect textDest = { Sint16 (position.x + 54), Sint16 (position.y + 25), 410, 30 };
 	const cPlayer* activePlayer = client->getActivePlayer();
 
-	size_t endIndex = std::min ( (index + 1u) * maxItems, activePlayer->savedReportsList.size());
+	size_t endIndex = std::min ( (index + 1u) * maxItems, (unsigned int)activePlayer->savedReportsList.size());
 	for (size_t i = index * maxItems; i != endIndex; ++i)
 	{
 		const sSavedReportMessage& savedReport = activePlayer->savedReportsList[i];
